Create sample version of SchemaTable component
Created a sample SchemaTable component using material-ui grid in
order to test how the component would look like. A simple JSON schema
example is statically embedded into the components for this purpose.
The result is that the two panels have each of their own
horizontal scroll, but < br /> tags must be embedded in the left
panel in order to align related rows with the right panel.