Add createNormalRow() and pushRow() method
Refactored renderDefault() method to create createNormalRow() and
pushRow() method. createNormalRow() creates an object containing a
NormalLeftRow and NormalRightRow component based on the given input
schema. pushRow() takes the resulting object of the former method and
pushes the components into the leftRows and rightRows respectively.
Refactoring the logic within renderDefault into separate methods
increases the reusability of creating rows and pushing them into the
arrays so that they can be easily accessible in future methods such as
renderArray and renderObject.