Fix same key error for object and array types
Fixed error for rendering the same key ('two children with the same key'
error in React). Problem was inherit in renderObject and renderArray in
that the each of the closingRows had the same key with the opening rows.
Fixed by creating the closingRows after all the array items or object
properties were pushed into the rows so that closingRows now have their
own unique key since they are not created at the same time with
openingRows.