Create Tree class for previewing schema structure
Created a Tree class to use as an intermediary data structure for
implementing ref expand/shrink feature.
The reason behind utilizing an intermediary structure is because
we need a structure to dynamically add and remove rows for the ref
type schemas. By using a tree structure, we can easily add and remove
nodes to represent subschemas more efficiently than simply using arrays.
Also, by containing ref schemas within nodes, we can also efficiently
store the state of expand/shrink of a ref row contained in its own unit.
The ultimate goal is to first create a tree structure based on the
schema and then create the schema table's rows based on that structure.