Change expand/shrink click method to ref row
Changed the ref's expand/shrink click method to apply to the entire ref
row instead of the ref icon button. This way, the user is able to click
the entire row to expand or collapse a ref. However, this only applies
to the first row of the ref when expanded.