Fix rows to display max 3 chips per line
Fixed NormalRightRow to display max of 3 chips (specification keywords)
per line. Also, matched NormalLeftRow to create blank line paddings to
match the number of lines with NormalRightRow.
The decision to have only a max of 3 chips per line instead of having an
automatic flex box is because creating matching lines between the left
and right panels would be complicated.