Add createTitleLine function in NormalRightRow
Added createTitleLine function to refactor title keyword from
createDescriptionLine. This way, only descriptions are displayed with
a tooltip using the OverflowLine component, while titles are simply
displayed with the Typography component. The result is that both title
and description are ellipsed when text is too long, but descriptions use
a tooltip (regardless of text-overflow) while titles do not.