Add forwardRef feature in Tooltip component
Add forwardRef to children components in Tooltip so that the tooltip
applies to even custom components. The result is that children nodes of
the Tooltip component are wrapped within a div and are passed properties
and ref of the material-ui's Tooltip so that a tooltip can apply to even
custom created components.