Refactor display to show the entire description
This switches from two columns with carefully-aligned heights of the
divs in those columns, to a float-based layout with two divs floated on
each row, separated by `clear: all`.  This avoids the need to balance
the number of fixed-height lines on each side, and thus allows
arbitrary-height descriptions with "normal" wrapping.  This is a big
usability improvement, as the descriptions are important content for
users!

This also adjusts the right row to always show the title first, then the
description, then the chips